expose more interfaces
improve name of exposed interfaces
