+++ tests/src/app/app.component.ts
@@ -1,10 +1,8 @@
 import { Component } from '@angular/core';
-import { Traverser } from 'angular-traversal';
 import { PloneViews, Services } from './lib';
 
 import { CustomViewView } from './custom';
-import { Status } from './lib/api.service';
-import { Authenticated } from '../../../src/authentication.service';
+import { AuthenticatedStatus, LoadingStatus } from './lib/interfaces';
 
 @Component({
   selector: 'app-root',
@@ -29,11 +27,11 @@ export class AppComponent {
 
   ngOnInit() {
     this.services.authentication.isAuthenticated
-      .subscribe((auth: Authenticated) => {
+      .subscribe((auth: AuthenticatedStatus) => {
         this.logged = auth.state;
       });
     this.services.api.status
-      .subscribe((status: Status) => {
+      .subscribe((status: LoadingStatus) => {
         this.loading = status.loading ? 'Loading...' : 'OK';
         this.error = status.error ? status.error.message : '';
       });
